requirements.txt
asgiref==3.5.2
asttokens==2.0.8
autopep8==1.7.0
backcall==0.2.0
bcrypt==3.2.2
boto3==1.24.80
botocore==1.27.80
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
colorama==0.4.5
cryptography==37.0.4
decorator==5.1.1
Django==4.1
django-console==0.4.7
django-cors-headers==3.13.0
django-environ==0.9.0
django-extensions==3.2.0
django-rest-framework==0.1.0
django-storages==1.13.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
executing==0.10.0
idna==3.3
ipython==8.4.0
jedi==0.18.1
jmespath==1.0.1
matplotlib-inline==0.1.5
paramiko==2.11.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.30
psycopg2==2.9.3
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
pycparser==2.21
Pygments==2.13.0
PyJWT==2.4.0
PyNaCl==1.5.0
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.4.0
toml==0.10.2
tornado==6.2
traitlets==5.3.0
tzdata==2022.2
urllib3==1.26.11
wcwidth==0.2.5
whitenoise==6.2.0
